Remove controller tests from system tests and paste into controller test spec
